+++ client/src/pages/Dashboard/CreateBoardForm.jsx
@@ -1,8 +1,8 @@
 import {useState, useContext} from 'react';
-import dataService from '../services/dataService';
-import {ModalContext} from '../contexts/ModalContext/ModalContext';
-import {useRoutingContext} from '../contexts/RoutingContext/routingContext';
-import {useSelectedBoardContext} from '../contexts/BoardContext/boardContext';
+import dataService from '../../services/dataService';
+import {ModalContext} from '../../contexts/ModalContext/ModalContext';
+import {useRoutingContext} from '../../contexts/RoutingContext/routingContext';
+import {useSelectedBoardContext} from '../../contexts/BoardContext/boardContext';
 
 const BoardForm = () => {
     const {handleModal, isModalOpen, handleClose, handleOpen} =
@@ -81,7 +81,7 @@ const BoardForm = () => {
                         </div>
 
                         <div>
-                            <div className='flex items-center justify-between'></div>
+                            <div className='flex items-center justify-between' />
                             <div>
                                 <label
                                     htmlFor='description'
@@ -97,7 +97,7 @@ const BoardForm = () => {
                                         onChange={handleChange}
                                         required
                                         className='block w-full rounded-md border-0 py-1.5 text-gray-900 shadow-sm ring-1 ring-inset ring-gray-300 placeholder:text-gray-400 focus:ring-2 focus:ring-inset focus:ring-indigo-600 sm:text-sm sm:leading-6'
-                                    ></textarea>
+                                     />
                                 </div>
                                 <p className='mt-3 text-sm leading-6 text-gray-600'>
                                     Write a few sentences about your project.
