+++ client/src/pages/Workspace/Task/Assignee/AssigneesList.jsx
@@ -20,15 +20,13 @@ export default function AssigneesList({users, formData, onChange}) {
     console.log(formData);
     return (
         <div className='flex gap-x-4 flex-wrap '>
-            {users.map((user) => {
-                return (
+            {users.map((user) => (
                     <AssigneeCheckbox
                         user={user}
                         onChange={onChange}
                         formData={formData}
                     />
-                );
-            })}
+                ))}
         </div>
     );
 }
