+++ client/src/pages/Workspace/components/WorkspaceHeader.jsx
@@ -1,17 +1,17 @@
 import {useState, useContext} from 'react';
 import {MdDelete, MdModeEdit, MdFileCopy} from 'react-icons/md';
 import {ToastContainer, toast} from 'react-toastify';
-import {useRoutingContext} from '../contexts/RoutingContext/routingContext';
-import {useAuthContext} from '../contexts/AuthContext/authContext';
-import {useSelectedBoardContext} from '../contexts/BoardContext/boardContext';
-import dataService from '../services/dataService';
-import formatDate from '../utils/formatDate';
+import {useRoutingContext} from '../../../contexts/RoutingContext/routingContext';
+import {useAuthContext} from '../../../contexts/AuthContext/authContext';
+import {useSelectedBoardContext} from '../../../contexts/BoardContext/boardContext';
+import dataService from '../../../services/dataService';
+import formatDate from '../../../utils/formatDate';
 import 'react-toastify/dist/ReactToastify.css';
-import BoardConfirmDelete from './BoardConfirmDelete';
-import {ModalContext} from '../contexts/ModalContext/ModalContext';
-import Modal from '../components/Modal';
+import BoardConfirmDelete from '../../../features/BoardConfirmDelete';
+import {ModalContext} from '../../../contexts/ModalContext/ModalContext';
+import Modal from '../../../components/Modal';
 import useEditingMode from '../hooks/useEditingMode';
-import TextAreaEditor from '../components/TextAreaEditor';
+import TextAreaEditor from '../../../components/TextAreaEditor';
 
 const WorkspaceHeader = ({boardInfo}) => {
     const {setCurrentPage} = useRoutingContext();
